requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
flask==3.1.0
fonttools==4.55.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sklearn==0.0
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.11
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing-extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.32
ultralytics-thop==2.0.11
urllib3==2.2.3
werkzeug==3.1.3
zipp==3.21.0
